chemin + controller ajouter et modifier

+++ src/main/java/org/example/spring_meuble/dao/CartItemRepository.java
@@ -3,7 +3,7 @@ package org.example.spring_meuble.dao;
 import org.example.spring_meuble.entity.CartItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
+public interface CartItemRepository extends JpaRepository<CartItem, Long> {
 
 
 }
